Beautify Whonix Welcome page in Tor Browser
The 7.0.9 release broke the Whonix welcome page in Tor Browser (https://forums.whonix.org/t/the-7-0-9-release-broke-the-whonix-landing-page-in-tor-browser/4489). Although the prvious welcome page still works, it is better to beautify it. This fix include replacing Whonix logo png with Ascii Art, removal of invlalid links and enlarging font size. Still this is a temporary workaround and please feel free to checkout to the previous one, once the breakage problem is fixed by TPO.